+++ stroom-integrationtest/src/test/java/stroom/pipeline/server/factory/MockPipelineElementRegistryFactory.java
@@ -116,7 +116,7 @@ public class MockPipelineElementRegistryFactory implements ElementRegistryFactor
                 return (T) new TextWriter();
             }
             if (elementClass.equals(StreamAppender.class)) {
-                return (T) new StreamAppender(null, null, null, null, null, null, null, null);
+                return (T) new StreamAppender(null, null, null, null, null, null, null, null, null);
             }
             if (elementClass.equals(FileAppender.class)) {
                 return (T) new FileAppender(null, null);