+++ stroom-pipeline/src/main/java/stroom/pipeline/server/writer/StreamAppender.java
@@ -27,9 +27,12 @@ import stroom.pipeline.server.errorhandler.ProcessException;
 import stroom.pipeline.server.factory.ConfigurableElement;
 import stroom.pipeline.server.factory.ElementIcons;
 import stroom.pipeline.server.factory.PipelineProperty;
+import stroom.pipeline.server.task.ProcessStatisticsFactory;
+import stroom.pipeline.server.task.ProcessStatisticsFactory.ProcessStatistics;
 import stroom.pipeline.shared.data.PipelineElementType;
 import stroom.pipeline.shared.data.PipelineElementType.Category;
 import stroom.pipeline.state.MetaData;
+import stroom.pipeline.state.RecordCount;
 import stroom.pipeline.state.StreamHolder;
 import stroom.pipeline.state.StreamProcessorHolder;
 import stroom.streamstore.server.StreamStore;
@@ -59,6 +62,7 @@ public class StreamAppender extends AbstractAppender {
     private final StreamTypeService streamTypeService;
     private final StreamProcessorHolder streamProcessorHolder;
     private final MetaData metaData;
+    private final RecordCount recordCount;
     private final StreamCloser streamCloser;
 
     private Feed feed;
@@ -68,6 +72,8 @@ public class StreamAppender extends AbstractAppender {
     private WrappedSegmentOutputStream wrappedSegmentOutputStream;
     private ByteCountOutputStream byteCountOutputStream;
 
+    private ProcessStatistics lastProcessStatistics;
+
     @Inject
     public StreamAppender(final ErrorReceiverProxy errorReceiverProxy,
                           final StreamStore streamStore,
@@ -76,6 +82,7 @@ public class StreamAppender extends AbstractAppender {
                           final StreamTypeService streamTypeService,
                           final StreamProcessorHolder streamProcessorHolder,
                           final MetaData metaData,
+                          final RecordCount recordCount,
                           final StreamCloser streamCloser) {
         super(errorReceiverProxy);
         this.errorReceiverProxy = errorReceiverProxy;
@@ -85,6 +92,7 @@ public class StreamAppender extends AbstractAppender {
         this.streamTypeService = streamTypeService;
         this.streamProcessorHolder = streamProcessorHolder;
         this.metaData = metaData;
+        this.recordCount = recordCount;
         this.streamCloser = streamCloser;
     }
 
@@ -149,9 +157,19 @@ public class StreamAppender extends AbstractAppender {
     private void close() {
         // Only do something if an output stream was used.
         if (streamTarget != null) {
-            // Write meta data.
-            final MetaMap metaMap = metaData.getMetaMap();
-            streamTarget.getAttributeMap().putAll(metaMap);
+            // Write process meta data.
+            streamTarget.getAttributeMap().putAll(metaData.getMetaMap());
+
+            // Get current process statistics
+            final ProcessStatistics processStatistics = ProcessStatisticsFactory.create(recordCount, errorReceiverProxy);
+            // Diff the current statistics with the last captured statistics.
+            final ProcessStatistics currentStatistics = processStatistics.substract(lastProcessStatistics);
+            // Set the last statistics.
+            lastProcessStatistics = processStatistics;
+
+            // Write statistics meta data.
+            currentStatistics.write(streamTarget.getAttributeMap());
+
             // We leave the streamCloser to close the stream target as it may
             // want to delete it instead
         }